+++ .github/workflows/backend-unit-test.yml
@@ -7,7 +7,7 @@ on:
     branches: [ "master" ]
 
 jobs:
-  http-clients-common-unit-test:
+  http-clients-common-test:
     name: HttpClients.Common Unit Test
     runs-on: ubuntu-latest
     env:
@@ -29,7 +29,7 @@ jobs:
     - name: Test
       run: dotnet test $ProjectName --no-build --verbosity normal
 
-  http-clients-unit-test:
+  http-clients-test:
     name: HttpClients Unit Test
     runs-on: ubuntu-latest
     env:
@@ -51,11 +51,11 @@ jobs:
     - name: Test
       run: dotnet test $ProjectName --no-build --verbosity normal
   
-  resource-web-application-unit-test:
-    name: ResourceWebApplication Unit Test
+  resource-web-application-controller-test:
+    name: ResourceWebApplication Controller Test
     runs-on: ubuntu-latest
     env:
-      ProjectName: HeadHunter.ResourceWebApplication.Tests
+      ProjectName: HeadHunter.ResourceWebApplication.ControllerTests
     steps:
     - uses: actions/checkout@v3
     - name: Setup .NET
