+++ HeadHunter.ResourceWebApplication.ControllerTests/Areas/HeadHunter/Controllers/VacanciesControllerTests.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Mvc;
 using Moq;
 using System.Net;
 
-namespace HeadHunter.ResourceWebApplication.Tests.Areas.HeadHunter.Controllers
+namespace HeadHunter.ResourceWebApplication.ControllerTests.Areas.HeadHunter.Controllers
 {
     public class VacanciesControllerTests
     {
