+++ HeadHunter.ResourceWebApplication.ControllerTests/HeadHunter.ResourceWebApplication.ControllerTests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
